% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing-methods.R
\name{canonicalize_by_subset}
\alias{canonicalize_by_subset}
\title{Return single contig for each cell based on filtering and sorting}
\usage{
canonicalize_by_subset(ccdb, ..., tie_break_keys = c("umis", "reads"),
  contig_fields = tie_break_keys, order = 1)
}
\arguments{
\item{ccdb}{`ContigCellDB`}

\item{...}{unquoted expressions passed to `dplyr::filter` that will be applied to the `contig_tbl`}

\item{tie_break_keys}{columns used to sort the `contig_tbl` in **decreasing** order}

\item{contig_fields}{columns in the `contig_tbl` that will be copied over to the `cell_tbl`}

\item{order}{`integer` specifying which entry (ordinal) will be selected from the sorted contig_tbl, within each cell}
}
\value{
`ContigCellDB` with additional fields in `cell_tbl`
}
\description{
Return single contig for each cell based on filtering and sorting
}
\examples{
# Report beta chain with highest umi-count, breaking ties with reads
beta = canonicalize_by_subset(ccdb_ex, chain == 'TRB',
tie_break_keys = c('umis', 'reads'),
contig_fields = c('umis', 'reads', 'chain', 'v_gene', 'd_gene', 'j_gene'))
head(beta$cell_tbl)
# Only adds fields to `cell_tbl`
stopifnot(all.equal(beta$cell_tbl[ccdb_ex$cell_pk],
ccdb_ex$cell_tbl[ccdb_ex$cell_pk]))
#Report cdr3 with highest UMI count, but only when > 5 UMIs support it
umi5 = canonicalize_by_subset(ccdb_ex, umis > 5,
tie_break_keys = c('umis', 'reads'), contig_fields = c('umis', 'cdr3'))
stopifnot(all(umi5$cell_tbl$umis > 5, na.rm = TRUE))
}
