% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigCellDB-methods.R
\docType{methods}
\name{$<-,ContigCellDB-method}
\alias{$<-,ContigCellDB-method}
\title{Creat a function of ContigCellDB object to replace values of its slots}
\usage{
\S4method{$}{ContigCellDB}(x, name) <- value
}
\arguments{
\item{x}{A ContigCellDB object}

\item{name}{Name of a slot for a ContigCellDB object}

\item{value}{The value assigned to a slot of ContigCellDB object}
}
\value{
A ContigCellDB object
}
\description{
Creat a function of ContigCellDB object to replace values of its slots
}
\examples{
ccdb_ex$contig_pk <- c("pop","barcode","contig_id")
}
