% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccdb_ex}
\alias{ccdb_ex}
\title{A preconstructed \code{ContigClusterDB} from the \code{contigs_qc} data}
\format{
\code{ccdb_ex = ContigCellDB_10XVDJ(contigs_qc, contig_pk = c('pop',   'sample', 'barcode', 'contig_id'), cell_pk = c('pop',   'sample', 'barcode'))}
}
\usage{
ccdb_ex
}
\description{
A preconstructed \code{ContigClusterDB} from the \code{contigs_qc} data
}
\seealso{
\link{contigs_qc}
}
\keyword{datasets}
