% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdb_join.R
\name{ccdb_join}
\alias{ccdb_join}
\title{Join dataframe or SingleCellExperiment object with ContigCellDB object}
\usage{
ccdb_join(template, ccdb, join_fun = dplyr::left_join, by = ccdb$cell_pk)
}
\arguments{
\item{template}{Dataframe or SingleCellExperiment object to be joined with ccdb.}

\item{ccdb}{A ContigCellDB object.}

\item{join_fun}{Function used for the join operation.}

\item{by}{A character vector of variables to join by.}
}
\value{
\code{\link[=ContigCellDB]{ContigCellDB()}}
}
\description{
Join dataframe or SingleCellExperiment object with ContigCellDB object
}
