% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-methods.R
\name{cland}
\alias{cland}
\title{Cluster "And" intersection}
\usage{
cland(X, Y, new_pk)
}
\arguments{
\item{X}{\code{ContigCellDB}}

\item{Y}{\code{ContigCellDB}}

\item{new_pk}{optional \code{character} naming the new pk.}
}
\description{
For each contig present in both \code{X} and \code{Y},
a new cluster is defined that combines cluster identities in both \code{X} and \code{Y}.
In the resulting \code{ContigCellDB}, two contigs are in the same cluster if they are in the
same cluster in \code{X} and the same cluster
in \code{Y}.  \code{X} and \code{Y} must have matching \code{contig_pk}.
The \code{contig_tbl} has fields from \code{X} for contigs present in both \code{X} and \code{Y}.
The  \code{cell_tbl} from \code{X} is carried forward unchanged, while the \code{cluster_tbl}
in the result contains the mapping between the ancestral clustering, and the derived.
}
\examples{
data(ccdb_ex)
ccdb_germ = cluster_germline(ccdb_ex, cluster_pk = 'germline_idx')
ccdb_cdr3 = cdhit_ccdb(ccdb_ex, 'cdr3_nt', type = 'DNA',
cluster_name = 'DNA97', identity = .965, min_length = 12, G = 1)
ccdb_cdr3 = cland(ccdb_cdr3, ccdb_germ)

}
