% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-testing.R
\name{cluster_filterset}
\alias{cluster_filterset}
\title{A filtration of clusters}
\usage{
cluster_filterset(min_number = 0, min_freq = 0, white_list = NULL)
}
\arguments{
\item{min_number}{\code{integer} At least this many cells}

\item{min_freq}{\code{numeric} At least this frequency}

\item{white_list}{\code{data.frame} keyed by cluster_pk that must match}
}
\value{
object representing the filtration (currently a list)
}
\description{
Return clusters that match all provided conditions
}
\examples{
cluster_filterset(min_number = 1, min_freq = 0)
}
