% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-methods.R
\name{cluster_germline}
\alias{cluster_germline}
\title{Cluster contigs by germline properties}
\usage{
cluster_germline(ccdb, segment_keys = c("v_gene", "j_gene", "chain"),
  cluster_name = "cluster_idx")
}
\arguments{
\item{ccdb}{`ContigCellDB`}

\item{segment_keys}{fields in `contig_tbl` that identify a cluster}

\item{cluster_name}{name of cluster to be added to `cluster_tbl`}
}
\value{
`ContigCellDB`
}
\description{
Cluster contigs by germline properties
}
\examples{
ccdb_ex = cluster_germline(ccdb_ex)
ccdb_ex$cluster_tbl
}
