% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-testing.R
\name{cluster_logistic_test}
\alias{cluster_logistic_test}
\title{Test clusters for differential usage}
\usage{
cluster_logistic_test(formula, ccdb, replicate_keys, cluster_whitelist,
  contig_filter_args = TRUE, tie_break_keys = c("umis", "reads"),
  keep_fit = FALSE, fitter = glm_glmer)
}
\arguments{
\item{formula}{the \strong{right-hand side} of a glmer or glm-style formula.}

\item{ccdb}{\code{\link[=ContigCellDB]{ContigCellDB()}}}

\item{replicate_keys}{keys naming columns in \code{ccdb$cell_tbl}}

\item{cluster_whitelist}{a table, keyed by \code{ccdb$cluster_pk} specifying the clusters to test. It does not alter which cells are included or how  canonicalization is performed. If omitted, all clusters will be tested and reported.}

\item{contig_filter_args}{an expression passed to \code{\link[dplyr:filter]{dplyr::filter()}}.
Unlike \code{filter}, multiple criteria must be \code{&} together, rather than using
commas to separate. These act on \code{ccdb$contig_tbl}}

\item{tie_break_keys}{(optional) \code{character} naming fields in \code{contig_tbl}
that are used sort the contig table in descending order.
Used to break ties if \code{contig_filter_args} does not return a unique contig
for each cluster}

\item{keep_fit}{\code{logical} as to whether the fit objects should be returned as a list column}

\item{fitter}{a function taking arguments \code{formula}, \code{data}, \code{is_mixed} and \code{keep_fit} that is run on each cluster.  Should return a \code{tibble} or \code{data.frame}}
}
\value{
table with one row per cluster/term.
}
\description{
Test clusters for differential usage
}
\examples{
data(ccdb_ex)
ccdb_ex = cluster_germline(ccdb_ex)
trav1 = dplyr::filter(ccdb_ex$cluster_tbl, v_gene == 'TRAV1')
cluster_logistic_test(~pop + (1|sample), ccdb_ex, 'sample', cluster_whitelist = trav1)
}
