% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immcantation-utils.R
\name{concat_fasta}
\alias{concat_fasta}
\title{Concatenate fasta files and rename sequence headers}
\usage{
concat_fasta(
  file_tbl,
  file_field,
  id_field,
  output_name = tempfile(pattern = "contigs", tmpdir = getwd()),
  sequence_name_prefix = "contig_id"
)
}
\arguments{
\item{file_tbl}{optional \code{tibble()} mapping between filenames and name}

\item{file_field}{column in \code{file_tbl} containing the input filenames}

\item{id_field}{column in \code{file_tbl} containing the identifier}

\item{output_name}{name for output fastas}

\item{sequence_name_prefix}{??}
}
\description{
This facilitates submission to that they can be submitted to highV-quest enmass.
}
