% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigCellDB-methods.R
\name{ContigCellDB}
\alias{ContigCellDB}
\alias{ContigCellDB_10XVDJ}
\title{Construct a ContigCellDB}
\usage{
ContigCellDB(contig_tbl, contig_pk, cell_tbl, cell_pk, cluster_tbl,
  cluster_pk = character())

ContigCellDB_10XVDJ(contig_tbl, contig_pk = c("barcode", "contig_id"),
  cell_pk = "barcode")
}
\arguments{
\item{contig_tbl}{a data frame of contigs, and additional fields describing their properties}

\item{contig_pk}{character vector naming fields in `contig_tbl` that uniquely identify a row/contig}

\item{cell_tbl}{a data frame of cell barcodes, and (optional) additional fields describing their properties}

\item{cell_pk}{character vector naming fields in `cell_tbl` that uniquely identify a cell barcode}

\item{cluster_tbl}{A data frame that provide cluster assignments for each contig}

\item{cluster_pk}{If `cluster_tbl` was provided, a list of character vector naming fields in `cluster_tbl` that uniquely identify a cluster}
}
\value{
\code{ContigCellDB}
}
\description{
Construct a ContigCellDB
}
\section{Functions}{
\itemize{
\item \code{ContigCellDB_10XVDJ}: provide defaults that correspond to identifiers in 10X VDJ data
}}

\examples{
data(contigs_qc)
ContigCellDB(contigs_qc, contig_pk = c('barcode', 'pop', 'sample', 'contig_id'), cell_pk = c('barcode', 'pop', 'sample'))
}
