% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigCellDB-methods.R
\docType{methods}
\name{rbind,ContigCellDB-method}
\alias{rbind,ContigCellDB-method}
\title{S4 Generics}
\usage{
\S4method{rbind}{ContigCellDB}(..., deparse.level = 1)
}
\arguments{
\item{...}{\code{\link[=ContigCellDB]{ContigCellDB()}}}

\item{deparse.level}{ignored}
}
\description{
S4 Generics
}
\examples{
data(ccdb_ex)
splat = split_cdb(ccdb_ex, 'chain', 'contig_tbl')
unite = rbind(splat$TRA, splat$TRB)
stopifnot(all.equal(unite, ccdb_ex))
}
