% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing-methods.R
\name{cross_tab_tbl}
\alias{cross_tab_tbl}
\title{Generate a 2d cross tab using arbitrary numbers of columns as factors}
\usage{
cross_tab_tbl(tbl, x_fields, y_fields)
}
\arguments{
\item{tbl}{\code{data.frame}}

\item{x_fields}{\code{character} fields in \code{tbl}}

\item{y_fields}{\code{character} fields in \code{tbl}}
}
\value{
\code{tibble}
}
\description{
As many rows as unique combs of x_fields
As many columns as unique combs of y_fields
No NA.
}
\examples{
cross_tab_tbl(mtcars, c('cyl', 'gear'), 'carb')
}
