% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{crosstab_by_celltype}
\alias{crosstab_by_celltype}
\title{make the cross table by celltype}
\usage{
crosstab_by_celltype(ccdb)
}
\arguments{
\item{ccdb}{A ContigCellDB object}
}
\value{
a cross table
}
\description{
make the cross table by celltype
}
\examples{
cdb = ContigCellDB(all_anno,contig_pk = c('barcode','pop','sample','contig_id'),cell_pk = c('barcode','pop','sample'))
total_umi <- crosstab_by_celltype(cdb)
}
