% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{filter_cdb}
\alias{filter_cdb}
\title{Filtering rows of tables in ContigCellDB object}
\usage{
filter_cdb(ccdb, ..., which_tbl = "contig_tbl")
}
\arguments{
\item{ccdb}{A ContigCellDB object}

\item{...}{conditions used to filter out rows}

\item{which_tbl}{name of the table needs to be filtered}
}
\value{
ContigCellDB object with updated table
}
\description{
Filtering rows of tables in ContigCellDB object
}
\examples{
data(contigs_qc)
MIN_CDR3_AA <- 6

cdb <- ContigCellDB_10XVDJ(contigs_qc, contig_pk = c('barcode', 'pop', 'sample', 'contig_id'), cell_pk = c('barcode', 'pop', 'sample'))
new_cdb <- filter_cdb(cdb,full_length, productive == 'True', high_confidence, chain != 'Multi', str_length(cdr3) > MIN_CDR3_AA)
}
