% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-methods.R
\name{fine_clustering}
\alias{fine_clustering}
\title{Perform additional clustering of sequences within groups}
\usage{
fine_clustering(clustering, sequence_key = "seq",
  type = clustering$type, max_affinity = NULL,
  keep_clustering_details = FALSE, ...)
}
\arguments{
\item{clustering}{`Clustering` object}

\item{sequence_key}{`character` naming column in `contig_tbl` with sequence}

\item{type}{'AA' or 'DNA'}

\item{max_affinity}{`numeric` naming the maximal affinity for the sparse affinity matrix that is constructed.  Not currently used.}

\item{keep_clustering_details}{`logical` -- should output of `fine_cluster_seqs` be kept as a list column}

\item{...}{passed to `fine_cluster_seqs`}
}
\value{
`Clustering` object with updated `contig_tbl` and `cluster_tbl`
}
\description{
Perform additional clustering of sequences within groups
}
