% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing-methods.R
\name{get_canonical_representative}
\alias{get_canonical_representative}
\title{Given a family of similar sequences, return a "representative"}
\usage{
get_canonical_representative(seqs, medoid_idx, warn_if_distinct = FALSE)
}
\arguments{
\item{seqs}{character vector}

\item{medoid_idx}{optional index into seqs}

\item{warn_if_distinct}{Should a warning be emitted if there are distinct elements in seqs?

If medoid_idx is supplied, the medoid sequence is returned, otherwise the longest
sequence is returned}
}
\value{
character vector
}
\description{
Given a family of similar sequences, return a "representative"
}
\examples{
get_canonical_representative(c('apple', 'manzana', 'pomme'))
}
