% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{guess_celltype}
\alias{guess_celltype}
\title{Guess the cell type of a contig from the chain ID}
\usage{
guess_celltype(chain)
}
\arguments{
\item{chain}{\code{character} which will be parsed to try to infer celltype}
}
\value{
contig table with \code{celltype} column
}
\description{
This function is likely dependent on annotations from 10X and change or break
as their pipeline changes.
}
\seealso{
\code{\link[=crosstab_by_celltype]{crosstab_by_celltype()}}
}
