% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-methods.R
\name{right_join_warn}
\alias{right_join_warn}
\alias{left_join_warn}
\title{Perform a \code{dplyr::left_join()} but check for non-key overlapping fields}
\usage{
right_join_warn(...)

left_join_warn(x, y, by, overwrite = FALSE, join = left_join, ...)
}
\arguments{
\item{...}{passed to joining function}

\item{x}{tbls to join}

\item{y}{tbls to join}

\item{by}{\code{character} specifying columns in \code{x} and \code{y} to key on.}

\item{overwrite}{\code{logical} -- should non-key fields in y be overwritten using x, or should a suffix (".y") be added}

\item{join}{function giving the type of join to perform, eg, left, right, inner, outer.}
}
\value{
\code{data.frame} or \code{tibble}
}
\description{
Perform a \code{dplyr} join, but either warn if the two tables
share non-key fields  If \code{overwrite = TRUE}, then shared columns will pull from \code{x} otherwise a suffix will be added to \code{y}.  To perform this check, \code{by} must be specified, and it is an error if it is not.
}
\section{Functions}{
\itemize{
\item \code{right_join_warn}: perform a \code{dplyr::right_join()}
}}

\examples{
left_join_warn(mtcars, mtcars, by  = 'mpg')
left_join_warn(mtcars, mtcars, by = 'mpg', overwrite = TRUE)
}
