% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigCellDB-methods.R
\name{filter_cdb}
\alias{filter_cdb}
\alias{mutate_cdb}
\title{Make changes (ceate new or update existed) on columns of tables in ContigCellDB object}
\usage{
filter_cdb(ccdb, ..., tbl = "contig_tbl")

mutate_cdb(ccdb, ..., tbl = "contig_tbl")
}
\arguments{
\item{ccdb}{A ContigCellDB object}

\item{...}{name and value pair of column that will be updated}

\item{tbl}{name of the table needs to update columns}
}
\value{
ContigCellDB object with updated table
}
\description{
Make changes (ceate new or update existed) on columns of tables in ContigCellDB object
}
\section{Functions}{
\itemize{
\item \code{filter_cdb}: Filter rows of a table in a \code{ContigCellDB} object
}}

\examples{
data(ccdb_ex)
subset_contig = filter_cdb(ccdb_ex,full_length, productive == 'True',
high_confidence, chain != 'Multi', nchar(cdr3) > 5)
subset_cell = filter_cdb(ccdb_ex, sample == 4, tbl = 'cell_tbl')
data(ccdb_ex)
new_contig = mutate_cdb(ccdb_ex, new_col = 1)
new_cell = mutate_cdb(ccdb_ex, new_col = 1, tbl = 'contig_tbl')
}
