% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustering.R
\name{plot_clustering}
\alias{plot_clustering}
\title{Visualization of clustering characteristics for a ContigCellDB object with inherited fine_clustering object}
\usage{
plot_clustering(
  ccdb,
  cluster_factors,
  type = c("heatmap", "network"),
  statistic = c("pearson", "contigs"),
  ncluster,
  chaintype
)
}
\arguments{
\item{ccdb}{A ContigCellDB object.}

\item{cluster_factors}{A character vector of variables in ccdb$contig table to visualize clusters by}

\item{type}{Type of visualization, a heatmap or a node-edge network plot}

\item{statistic}{Cluster characteristics visualized by pearson residuals or raw contig counts}

\item{ncluster}{Optional integer. Only plot characteristics that have at least nclusters (for clarity of visualization)}

\item{chaintype}{Character in ccdb$contig_tbl$chain. If passed will subset contigs belonging to specified chain (IGH,IGK,IGL,TRA,TRB)}
}
\value{
A ggraph object if type == 'network', and a ggplot object if type == 'heatmap'
}
\description{
Visualization of clustering characteristics for a ContigCellDB object with inherited fine_clustering object
}
\examples{
data(ccdb_ex)
ccdb_germline_ex = cluster_germline(ccdb_ex, segment_keys = c('v_gene', 'j_gene', 'chain'), cluster_pk = 'segment_idx')
ccdb_germline_ex = fine_clustering(germline_cluster, sequence_key = 'cdr3_nt', type = 'DNA')
plot_clustering(ccdb_germline_ex,cluster_factors = c('v_gene','j_gene'), statistic = 'pearson', type = 'network' ,ncluster = 50, chaintype = 'IGH')
}
