% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation-testing.R
\name{purity}
\alias{purity}
\title{Calculate number of cluster-subject singletons for the purposes of permutation testing}
\usage{
purity(cluster_idx, subject)
}
\arguments{
\item{cluster_idx}{factor-like cluster variable}

\item{subject}{factor-like subject}
}
\value{
average number of singletons
}
\description{
Calculate number of cluster-subject singletons for the purposes of permutation testing
}
\examples{
message("see example(cluster_permute_test)")
}
\seealso{
\code{\link[=cluster_permute_test]{cluster_permute_test()}}
}
