% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdhit-methods.R
\name{query_cdhit}
\alias{query_cdhit}
\title{Query the web interface to cdhit}
\usage{
query_cdhit(sequences, results_path = NULL, identity_cutoff = 0.9,
  bandwidth = 20, results_timeout = 120)
}
\arguments{
\item{sequences}{An object of class \code{AAStringSet}}

\item{results_path}{If non-null, a \code{character} specifying a path to a results.  Directories therein must exist. If null, the R temporary directory will be used.}

\item{identity_cutoff}{minimum identity to be clustered together}

\item{bandwidth}{see CDhit docs}

\item{results_timeout}{number of seconds to wait for result}
}
\value{
path to results file
}
\description{
This sends a query to \url{http://weizhong-lab.ucsd.edu/cdhit-web-server/cgi-bin/index.cgi?cmd=cd-hit}
and returns a link to results.
}
\examples{
\dontrun{
fasta_path = system.file('extdata', 'demo.fasta', package='CellaRepertorium')
aaseq = Biostrings::readAAStringSet(fasta_path)
results_path = query_cdhit(aaseq[1:100])
read_cdhit(results_path)
}
}
\seealso{
AAStringSet
}
