% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdhit-methods.R
\name{read_cdhit}
\alias{read_cdhit}
\title{Parse a CD-hit results file as a data.frame}
\usage{
read_cdhit(file)
}
\arguments{
\item{file}{path to a file with cdhit results}
}
\value{
\code{data.frame} with columns
}
\description{
Parses a cd-hit (\url{http://weizhong-cluster.ucsd.edu/cdhit_suite/cgi-bin/index.cgi?cmd=cd-hit}) cluster file and returns it as a \code{data.frame}.
The \code{data.frame} contains the columns `cluster_idx`: the cluster number, `query_idx`: the fasta identifier of the query sequence, `homology_pct`: the percent similarity to the exemplar sequence, `member_idx`, an arbitrary enumeration of each sequence in the a cluster, `len`: the length of the query sequence.
}
\examples{
demo_path = system.file('extdata', 'demo150.clstr.sorted', package='CellaRepertorium')
out = read_cdhit(demo_path)
}
