% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigCellDB-methods.R
\name{[[,ContigCellDB,character,missing-method}
\alias{[[,ContigCellDB,character,missing-method}
\alias{[,ContigCellDB,ANY,missing,ANY-method}
\alias{[,ContigCellDB,ANY,missing-method}
\alias{dim,ContigCellDB-method}
\alias{dimnames,ContigCellDB-method}
\alias{nrow,ContigCellDB-method}
\alias{ncol,ContigCellDB-method}
\title{\code{data.frame}-like mutation/accessor generics for \code{ContigCellDB} objects}
\usage{
\S4method{[[}{ContigCellDB,character,missing}(x, i, j, ...)

\S4method{[}{ContigCellDB,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{dim}{ContigCellDB}(x)

\S4method{dimnames}{ContigCellDB}(x)

\S4method{nrow}{ContigCellDB}(x)

\S4method{ncol}{ContigCellDB}(x)
}
\arguments{
\item{x}{\code{ContigCellDB}}

\item{i}{integer or character index}

\item{j}{ignored}

\item{...}{ignored}

\item{drop}{ignored}
}
\value{
See details.
}
\description{
A \code{ContigCellDB} pretend to be a \code{cell_tbl} data.frame in several regards.
This is to enable nesting \code{ContigCellDB} objects in the \code{colData} of a \code{SingleCellExperiment}
and so that various plotting functionality in \code{scater} can do something sensible.
}
\details{
If \code{x} a \code{ContigCellDB}, then \code{dim(x)} and \code{dimnames(x)} return \code{dim(x$cell_tbl)} and \code{dimnames(x$cell_tbl)}, respectively, and \code{x[[col]]}  returns \code{x$cell_tbl[[col]]}.
Likewise indexing with \code{x[i,]} returns cells indexed by \code{i}.
Finally \code{as.data.frame(x)} returns \code{x$cell_tbl}.
}
\examples{
 data(ccdb_ex)
 ccdb_ex[1:10,]
 head(ccdb_ex[['barcode']])
 dim(ccdb_ex)
 dimnames(ccdb_ex)
}
