% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigCellDB-methods.R
\docType{methods}
\name{$,ContigCellDB-method}
\alias{$,ContigCellDB-method}
\title{Creat a method for ContigCellDB object to access its slots}
\usage{
\S4method{$}{ContigCellDB}(x, name)
}
\arguments{
\item{x}{A ContigCellDB object}

\item{name}{a slot of a ContigCellDB object (one of  `c('contig_tbl', 'cell_tbl', 'contig_pk', 'cell_pk', 'cluster_tbl', 'cluster_pk')`)}
}
\value{
Slots of ContigCellDB
}
\description{
Creat a method for ContigCellDB object to access its slots
}
\examples{
ccdb_ex$contig_tbl
ccdb_ex$cell_tbl
ccdb_ex$cluster_tbl
}
