% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigCellDB-methods.R
\docType{methods}
\name{$,ContigCellDB-method}
\alias{$,ContigCellDB-method}
\title{Creat a method for ContigCellDB object to access its slots}
\usage{
\S4method{$}{ContigCellDB}(x, name)
}
\arguments{
\item{x}{A ContigCellDB object}

\item{name}{Name of a slot for a ContigCellDB object}
}
\value{
Slots of ContigCellDB
}
\description{
Creat a method for ContigCellDB object to access its slots
}
\examples{
ccdb_ex$cluster_tbl
}
