% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdhit-methods.R
\name{cdhit}
\alias{cdhit}
\title{CDHIT}
\usage{
cdhit(seqs, identity = 0.9, kmerSize = 5, name = "CD-Hit",
  showProgress = interactive(), ...)
}
\arguments{
\item{seqs}{\code{AAseq} or \code{DNAseq} (untested..)}

\item{identity}{minimum proportion identity}

\item{kmerSize}{word size.  5 for 70%-90% identity}

\item{name}{program name (?)}

\item{showProgress}{show a status bar}

\item{...}{other arguments passed to cdhit, see https://github.com/weizhongli/cdhit/wiki/3.-User's-Guide#CDHIT for details}
}
\value{
vector of \code{integer} of length \code{seqs} providing the cluster ID for each sequence
}
\description{
CDHIT
}
