% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdhit-methods.R
\name{cdhit_ccdb}
\alias{cdhit_ccdb}
\title{Use \code{\link[=cdhit]{cdhit()}} to cluster a \code{\link[=ContigCellDB]{ContigCellDB()}}}
\usage{
cdhit_ccdb(
  ccdb,
  sequence_key,
  type = c("DNA", "AA"),
  cluster_pk = "cluster_idx",
  ...
)
}
\arguments{
\item{ccdb}{An object of class \code{\link[=ContigCellDB]{ContigCellDB()}}}

\item{sequence_key}{\code{character} naming the column in the \code{contig_tbl} containing the sequence to be clustered}

\item{type}{one of 'DNA' or 'AA'}

\item{cluster_pk}{\code{character} specifying key, and name for the clustering.}

\item{...}{
  Arguments passed on to \code{\link[=cdhit]{cdhit}}
  \describe{
    \item{\code{identity}}{minimum proportion identity}
    \item{\code{kmerSize}}{word size.  If NULL, it will be chosen automatically based on
the identity.
You may need to lower it below 5 for AAseq with identity less than .7.}
    \item{\code{min_length}}{Minimum length for sequences to be clustered.  An error if
something smaller is passed.}
    \item{\code{s}}{fraction of shorter sequence covered by alignment.}
    \item{\code{showProgress}}{show a status bar}
  }}
}
\value{
\code{\link[=ContigCellDB]{ContigCellDB()}}
}
\description{
Use \code{\link[=cdhit]{cdhit()}} to cluster a \code{\link[=ContigCellDB]{ContigCellDB()}}
}
\examples{
data(ccdb_ex)
res = cdhit_ccdb(ccdb_ex, 'cdr3_nt', type = 'DNA',
cluster_name = 'DNA97', identity = .965, min_length = 12, G = 1)
res$cluster_tbl
res$contig_tbl
res$cluster_pk
}
\seealso{
\code{\link[=cdhit]{cdhit()}}
}
