% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{crosstab_by_celltype}
\alias{crosstab_by_celltype}
\title{Count contig UMIs by celltype}
\usage{
crosstab_by_celltype(ccdb)
}
\arguments{
\item{ccdb}{A ContigCellDB object}
}
\value{
a table, keyed by \code{cell_pk} counting UMIs per celltype
}
\description{
Count contig UMIs by celltype
}
\examples{
data(ccdb_ex)
nrow(ccdb_ex$cell_tbl)
total_umi = crosstab_by_celltype(ccdb_ex)
nrow(total_umi)
}
\seealso{
\code{\link[=guess_celltype]{guess_celltype()}}
}
