% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation-testing.R
\name{.cluster_permute_test}
\alias{.cluster_permute_test}
\title{Cell permutation tests (internal)}
\usage{
.cluster_permute_test(labels, covariates, statistic, n_perm, alternative,
  ...)
}
\arguments{
\item{labels}{\code{factor} of length n}

\item{covariates}{\code{factor} of length n}

\item{statistic}{function of label and covariate}

\item{n_perm}{number of permutations.}

\item{alternative}{\code{character} naming the direction \code{statistic} should be fall under the alternative hypothesis}

\item{...}{passed to \code{statistic}}
}
\description{
Cell permutation tests (internal)
}
