% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-methods.R
\name{fine_clustering}
\alias{fine_clustering}
\title{Perform additional clustering of sequences within groups}
\usage{
fine_clustering(ccdb, sequence_key, type, max_affinity = NULL,
  keep_clustering_details = FALSE, ...)
}
\arguments{
\item{ccdb}{A ContigCellDB object}

\item{sequence_key}{`character` naming column in `contig_tbl` with sequence}

\item{type}{'AA' or 'DNA'}

\item{max_affinity}{`numeric` naming the maximal affinity for the sparse affinity matrix that is constructed.  Not currently used.}

\item{keep_clustering_details}{`logical` -- should output of `fine_cluster_seqs` be kept as a list column}

\item{...}{passed to `fine_cluster_seqs`}
}
\value{
`ContigCellDB` object with updated `contig_tbl` and `cluster_tbl`
}
\description{
Perform additional clustering of sequences within groups
}
\examples{
library(dplyr)
ccdb_ex_small = ccdb_ex
ccdb_ex_small$cell_tbl = ccdb_ex_small$cell_tbl[1:200,]
ccdb_ex_small = cdhit_ccdb(ccdb_ex_small,
sequence_key = 'cdr3_nt', type = 'DNA', cluster_name = 'DNA97',
identity = .965, min_length = 12, G = 1)
ccdb_ex_small = fine_clustering(ccdb_ex_small, sequence_key = 'cdr3_nt', type = 'DNA')

# Canonicalize with the medoid contig is probably what is most common
ccdb_medoid = canonicalize_cluster(ccdb_ex_small)

# But there are other possibilities.
# To pass multiple "AND" filter arguments must use &
ccdb_umi = canonicalize_cluster(ccdb_ex_small,
contig_filter_args = chain == 'TRA' & length > 500, tie_break_keys = 'umis',
contig_fields = c('chain', 'length'))
ccdb_umi$cluster_tbl \%>\% dplyr::select(chain, length) \%>\% summary()
}
