% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigCellDB-methods.R
\name{filter_cdb}
\alias{filter_cdb}
\alias{mutate_cdb}
\title{Create new or update existing columns of \code{ContigCellDB} tables}
\usage{
filter_cdb(ccdb, ..., tbl = "contig_tbl")

mutate_cdb(ccdb, ..., tbl = "contig_tbl")
}
\arguments{
\item{ccdb}{\code{\link[=ContigCellDB]{ContigCellDB()}}}

\item{...}{name and value pair of column that will be updated}

\item{tbl}{\code{character.} One of \code{contig_tbl}, \code{cell_tbl} or \code{cluster_tbl}, naming the table to be updated.}
}
\value{
ContigCellDB object with updated table
}
\description{
Create new or update existing columns of \code{ContigCellDB} tables
}
\section{Functions}{
\itemize{
\item \code{filter_cdb}: Filter rows of a table in a \code{ContigCellDB} object
}}

\examples{
data(ccdb_ex)
subset_contig = filter_cdb(ccdb_ex,full_length, productive == 'True',
high_confidence, chain != 'Multi', nchar(cdr3) > 5)
subset_cell = filter_cdb(ccdb_ex, sample == 4, tbl = 'cell_tbl')
data(ccdb_ex)
new_contig = mutate_cdb(ccdb_ex, new_col = 1)
new_cell = mutate_cdb(ccdb_ex, new_col = 1, tbl = 'contig_tbl')
}
\seealso{
\code{\link[dplyr:mutate]{dplyr::mutate()}}

\code{\link[dplyr:filter]{dplyr::filter()}}
}
