% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{mutate_cdb}
\alias{mutate_cdb}
\title{Make changes (ceate new or update existed) on columns of tables in ContigCellDB object}
\usage{
mutate_cdb(ccdb, ..., which_tbl = "contig_tbl")
}
\arguments{
\item{ccdb}{A ContigCellDB object}

\item{...}{name and value pair of column that will be updated}

\item{which_tbl}{name of the table needs to update columns}
}
\value{
ContigCellDB object with updated table
}
\description{
Make changes (ceate new or update existed) on columns of tables in ContigCellDB object
}
\examples{
data(contigs_qc)
MIN_CDR3_AA <- 6
cdb <- ContigCellDB_10XVDJ(contigs_qc, contig_pk = c('barcode', 'pop', 'sample', 'contig_id'), cell_pk = c('barcode', 'pop', 'sample'))
new_cdb <- mutate_cdb(cdb, new_col = 1)
}
