% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation-testing.R
\name{plot_permute_test}
\alias{plot_permute_test}
\title{Plot of histogram of permuted test statistics and observed test statistic}
\usage{
plot_permute_test(perm_test, title = "Permuted and observed test statistics")
}
\arguments{
\item{perm_test}{\code{list} that has been returned from running \code{cluster_permute_test()}}
}
\value{
A ggplot2 visualization of permutation test statistics
}
\description{
Plot of histogram of permuted test statistics and observed test statistic
}
\examples{
# See cluster_permute_test to set up \code{ccdb_test}
clust_test = cluster_permute_test(ccdb_test, 'subject', 'cluster_idx',
statistic = purity, n_perm  = 50)

plot_permute_test(perm_test = clust_test)

}
