% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation-testing.R
\name{purity}
\alias{purity}
\title{Calculate number of cluster-subject singletons}
\usage{
purity(cluster_idx, subject)
}
\arguments{
\item{cluster_idx}{factor-like cluster variable}

\item{subject}{factor-like subject}
}
\value{
average number of singletons
}
\description{
Calculate number of cluster-subject singletons
}
\seealso{
cluster_permute_test
}
